getCohortCharacteristics {PioneerWatchfulWaiting}	R Documentation

Create characterization of a cohort

Description

Computes features using all drugs, conditions, procedures, etc. observed on or prior to the cohort index date.

Usage

getCohortCharacteristics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortId,
  covariateSettings = FeatureExtraction::createDefaultCovariateSettings()
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write privileges for storing temporary tables.

cohortDatabaseSchema
Schema name where your cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortTable
Name of the cohort table.

cohortId
The cohort definition ID used to reference the cohort in the cohort table.

covariateSettings
Either an object of type covariateSettings as created using one of the createCovariate functions in the FeatureExtraction package, or a list of such objects.

Value

A data frame with cohort characteristics.

[Package PioneerWatchfulWaiting version 0.4.4 Index]