getInclusionStatistics {PioneerWatchfulWaiting}	R Documentation

Get statistics on cohort inclusion criteria

Description

Get statistics on cohort inclusion criteria

Usage

getInclusionStatistics(
  connectionDetails = NULL,
  connection = NULL,
  resultsDatabaseSchema,
  cohortId,
  simplify = TRUE,
  cohortTable = "cohort",
  cohortInclusionTable = paste0(cohortTable, "_inclusion"),
  cohortInclusionResultTable = paste0(cohortTable, "_inclusion_result"),
  cohortInclusionStatsTable = paste0(cohortTable, "_inclusion_stats"),
  cohortSummaryStatsTable = paste0(cohortTable, "_summary_stats")
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

resultsDatabaseSchema
Schema name where the statistics tables reside. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortId
The cohort definition ID used to reference the cohort in the cohort table.

simplify
Simply output the attrition table?

cohortTable
Name of the cohort table. Used only to conveniently derive names of the four rule statistics tables.

cohortInclusionTable
Name of the inclusion table, one of the tables for storing inclusion rule statistics.

cohortInclusionResultTable
Name of the inclusion result table, one of the tables for storing inclusion rule statistics.

cohortInclusionStatsTable
Name of the inclusion stats table, one of the tables for storing inclusion rule statistics.

cohortSummaryStatsTable
Name of the summary stats table, one of the tables for storing inclusion rule statistics.

Value

If 'simplify = TRUE', this function returns a single data frame. Else a list of data frames is returned.

[Package PioneerWatchfulWaiting version 0.4.4 Index]