getInclusionStatisticsFromFiles {PioneerWatchfulWaiting}	R Documentation

Get inclusion criteria statistics from files

Description

Gets inclusion criteria statistics from files, as stored when using the ROhdsiWebApi::insertCohortDefinitionSetInPackage function with generateStats = TRUE.

Usage

getInclusionStatisticsFromFiles(
  cohortId,
  folder,
  cohortInclusionFile = file.path(folder, "cohortInclusion.csv"),
  cohortInclusionResultFile = file.path(folder, "cohortIncResult.csv"),
  cohortInclusionStatsFile = file.path(folder, "cohortIncStats.csv"),
  cohortSummaryStatsFile = file.path(folder, "cohortSummaryStats.csv"),
  simplify = TRUE
)

Arguments

cohortId
The cohort definition ID used to reference the cohort in the cohort table.

folder
The path to the folder where the inclusion statistics are stored.

cohortInclusionFile
Name of the inclusion table, one of the tables for storing inclusion rule statistics.

cohortInclusionResultFile
Name of the inclusion result table, one of the tables for storing inclusion rule statistics.

cohortInclusionStatsFile
Name of the inclusion stats table, one of the tables for storing inclusion rule statistics.

cohortSummaryStatsFile
Name of the summary stats table, one of the tables for storing inclusion rule statistics.

simplify
Simply output the attrition table?

Value

If 'simplify = TRUE', this function returns a single data frame. Else a list of data frames is returned.

[Package PioneerWatchfulWaiting version 0.4.4 Index]