preMergeResultsFiles {PioneerWatchfulWaiting}	R Documentation

Premerge Shiny results files

Description

If there are many results files, starting the Shiny app may take a very long time. This function already does most of the preprocessing, increasing loading speed.

The merged data will be stored in the same folder, and will automatically be recognized by the Shiny app.

Usage

preMergeResultsFiles(dataFolder)

Arguments

dataFolder
folder where the exported zip files for the diagnostics are stored. Use the runStudy function to generate these zip files. Zip files containing results from multiple databases can be placed in the same folder.

[Package PioneerWatchfulWaiting version 0.4.4 Index]