writeBulkCharacteristics {PioneerWatchfulWaiting}	R Documentation

Write cohort characteristics in bulk to the file system

Description

This function will retrieve the results from the temp tables created in createBulkCharacteristics

Usage

writeBulkCharacteristics(
  connection,
  oracleTempSchema,
  counts,
  minCellCount,
  databaseId,
  exportFolder
)

[Package PioneerWatchfulWaiting version 0.4.4 Index]